pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "solfoundry-cli"
version = "0.1.0"
description = "SolFoundry Bounty CLI - Interact with SolFoundry bounties from the terminal"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "SolFoundry Contributor", email = "contributor@solfoundry.io"}
]
requires-python = ">=3.8"
dependencies = [
    "typer>=0.9.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
sf = "solfoundry_cli.main:app"
solfoundry = "solfoundry_cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["solfoundry_cli*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
